pyproject.toml
[project]
name = "db-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.40",
]
